add link to Shiny simulation tutorial
The Tutorial is an interactive R Markdown document, currently hosted on shinyapps.io.  It aims to take the reader step-by-step through the construction of a reasonably full-featured simulation app that lets statistics students explore, through simulation, the coverage properties of the classical t-intervals for a population mean.  After completing the tutorial the reader will be able to write his/her own simulation apps---hopefully having been spared some of the struggle that I went through when I first learned Shiny in the Spring of 2014.